chore: stop using gulp as a test runner
