+++ package.json
@@ -41,7 +41,7 @@
     "typescript-eslint": "^8.66.0"
   },
   "scripts": {
-    "test": "npx nx run-many -t test --exclude=blockly && npx nx run blockly:test",
+    "test": "npx nx run-many -t test",
     "lint": "npm run lint --ws --if-present",
     "lint-fix": "npm run lint-fix --ws --if-present",
     "build": "npx nx run-many -t build --exclude=blockly-docs",
