+++ packages/blockly/package.json
@@ -52,12 +52,24 @@
       "test": {
         "dependsOn": [
           {
+            "projects": "blockly",
+            "target": "package"
+          },
+          {
             "projects": "@blockly/block-test",
             "target": "build"
           },
           {
             "projects": "@blockly/dev-tools",
             "target": "build"
+          }
+        ]
+      },
+      "test-compile-advanced-browser": {
+        "dependsOn": [
+          {
+            "projects": "blockly",
+            "target": "test-compile-advanced"
           }
         ]
       },
@@ -101,13 +113,16 @@
     "rebuildAdvancedPlayground": "npm run build --workspace=@blockly/dev-tools && node \"scripts/prepare_advanced_playground.mjs\"",
     "rebuildBlockly": "npm run build && node \"scripts/prepare_advanced_playground.mjs\"",
     "tsc": "gulp tsc",
-    "test": "gulp test",
+    "test": "nx run-many -p blockly --targets=lint,test-renamings,test-mocha-node,test-generators,test-type-definitions,test-compile-advanced,test-compile-advanced-browser",
     "test-browser": "npx mocha --config tests/browser/.mocharc.js",
     "test-mocha-node": "npm run test-mocha-typecheck && npx mocha --config tests/mocha/.mocharc.node.cjs \"tests/mocha/**/*_test.{js,ts}\"",
     "test-mocha-typecheck": "tsc -p tests/mocha/tsconfig.json",
-    "test-generators": "gulp testGenerators",
+    "test-generators": "node \"tests/generators/test_generators.mjs\"",
     "test-mocha-interactive": "npm run build && node scripts/prepare_mocha_bundle.mjs && concurrently -n tsc,tsc:tests,esbuild,python3 \"tsc --watch --preserveWatchOutput --outDir \"build/src\" --declarationDir \"build/declarations\"\" \"npm run test-mocha-typecheck -- --watch --preserveWatchOutput\" \"esbuild build/tests/bundle-entry.js --bundle --sourcemap --alias:blockly/core=./build/src/core/blockly.js --alias:blockly/blocks=./build/src/blocks/blocks.js --alias:blockly=./build/src/core/blockly.js --outfile=build/tests/mocha-bundle.js --servedir=. --serve=127.0.0.1:8080 --watch=forever --log-level=warning\" \"python3 -m webbrowser 'http://localhost:8080/tests/mocha/index.html'\"",
     "test-compile-advanced": "gulp buildAdvancedCompilationTest --debug",
+    "test-compile-advanced-browser": "node \"tests/compile/webdriver.js\"",
+    "test-renamings": "node \"tests/migration/validate-renamings.mjs\"",
+    "test-type-definitions": "tsc -p ./tests/typescript/tsconfig.json -outDir \"build/src\"",
     "updateGithubPages": "node scripts/update_github_pages.mjs --upstream",
     "updateGithubPages-staging": "node scripts/update_github_pages.mjs --use-local"
   },
